refactor: break flowmodel into separate files
