+++ nessai/flowmodel/__init__.py
@@ -2,7 +2,8 @@
 """
 Classes for interfacing with flows. This includes training and sampling.
 """
-from .base import FlowModel, update_config
+from .base import FlowModel
+from .utils import update_config
 
 __all__ = [
     "FlowModel",
